requirements.txt
attrs==21.2.0
black==21.10b0
bleach==4.1.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
coverage==6.1.2
docutils==0.18
ghp-import==2.0.2
idna==3.3
importlib-metadata==4.8.2
iniconfig==1.1.1
Jinja2==3.0.3
keyring==23.2.1
Markdown==3.3.6
MarkupSafe==2.0.1
mergedeep==1.3.4
mkdocs==1.2.3
mypy-extensions==0.4.3
packaging==21.2
pathspec==0.9.0
pkginfo==1.7.1
platformdirs==2.4.0
pluggy==1.0.0
py==1.11.0
py4j==0.10.9.2
pydantic==1.8.2
Pygments==2.10.0
pyparsing==2.4.7
pyspark==3.2.0
pytest==6.2.5
pytest-cov==3.0.0
pytest-env==0.6.2
pytest-mock==3.6.1
python-dateutil==2.8.2
PyYAML==6.0
pyyaml_env_tag==0.1
readme-renderer==30.0
regex==2021.11.10
requests==2.26.0
requests-mock==1.9.3
requests-toolbelt==0.9.1
rfc3986==1.5.0
setuptools-git-versioning==1.7.4
six==1.16.0
toml==0.10.2
tomli==1.2.2
tqdm==4.62.3
twine==3.6.0
typing_extensions==4.0.0
urllib3==1.26.7
watchdog==2.1.6
webencodings==0.5.1
-e git+ssh://git@github.com/semanticinsight/yetl-framework.git@acf6232cbec4a7f13f55ea79b047a4ab3d09ee42#egg=yetl_framework
zipp==3.6.0
